pyproject.toml
[tool.poetry]
name = "nonebot_plugin_cocdicer"
version = "0.2.2"
description = "A COC dice plugin for Nonebot2"
authors = ["abrahumlink <307887491@qq.com>"]
documentation = "https://github.com/abrahum/nonebot_plugin_cocdicer#readme"
homepage = "https://github.com/abrahum/nonebot_plugin_cocdicer"
keywords = ["nonebot", "nonebot2", "coc", "dice"]
license = "GPLv3"
readme = "README.md"
repository = "https://github.com/abrahum/nonebot_plugin_cocdicer"

[tool.poetry.dependencies]
python = "^3.7"
nonebot2 = "^2.0.0-alpha.10"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
